+++ maven2-plugin/src/main/java/org/sitemesh/maven2Plugin/Setting.java
@@ -1,4 +1,4 @@
-package org.sitemesh.sitemesh;
+package org.sitemesh.maven2Plugin;
 
 import java.io.File;
 import java.util.List;
